+++ Source/PropertyTools.Wpf.Tests/ItemsBag/ItemsBagTests.cs
@@ -24,16 +24,16 @@ namespace PropertyTools.Wpf.Tests
             var props = td.GetProperties();
 
             var p1 = props.Find("Checked", false);
-            Assert.AreEqual(false, p1.IsReadOnly);
+            Assert.That(p1.IsReadOnly, Is.EqualTo(false));
             p1.SetValue(bag, true);
-            Assert.AreEqual(true, t0.Checked);
-            Assert.AreEqual(true, t0.IsChecked);
+            Assert.That(t0.Checked, Is.EqualTo(true));
+            Assert.That(t0.IsChecked, Is.EqualTo(true));
 
             var p2 = props.Find("IsChecked", false);
 
-            Assert.AreEqual(true, p2.IsReadOnly);
+            Assert.That(p2.IsReadOnly, Is.EqualTo(true));
             Assert.Throws<ArgumentException>(() => p2.SetValue(bag, false));
-            Assert.AreEqual(true, t0.IsChecked);
+            Assert.That(t0.IsChecked, Is.EqualTo(true));
         }
 
         [Test]
@@ -45,7 +45,7 @@ namespace PropertyTools.Wpf.Tests
             var provider = new ItemsBagTypeDescriptionProvider();
             var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
             var p1 = td.GetProperties().Find("Checked", false);
-            Assert.AreEqual(null, p1.GetValue(bag));
+            Assert.That(p1.GetValue(bag), Is.EqualTo(null));
         }
 
         [Test]
@@ -57,7 +57,7 @@ namespace PropertyTools.Wpf.Tests
             var provider = new ItemsBagTypeDescriptionProvider();
             var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
             var p1 = td.GetProperties().Find("Checked", false);
-            Assert.AreEqual(true, p1.GetValue(bag));
+            Assert.That(p1.GetValue(bag), Is.EqualTo(true));
         }
 
         [Test]
@@ -69,7 +69,7 @@ namespace PropertyTools.Wpf.Tests
             var provider = new ItemsBagTypeDescriptionProvider();
             var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
             var p1 = td.GetProperties().Find("Name", false);
-            Assert.AreEqual(null, p1.GetValue(bag));
+            Assert.That(p1.GetValue(bag), Is.EqualTo(null));
         }
 
         [Test]
@@ -81,7 +81,7 @@ namespace PropertyTools.Wpf.Tests
             var provider = new ItemsBagTypeDescriptionProvider();
             var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
             var p1 = td.GetProperties().Find("Name", false);
-            Assert.AreEqual("John", p1.GetValue(bag));
+            Assert.That(p1.GetValue(bag), Is.EqualTo("John"));
         }
 
         private class TestObject